Move extended field fetching to new component
